parser: Add table@{PHYSICAL_CHECK} syntax
This adds a new index hint option to trigger the physical check code
during scanning. It is an index hint because the check pertains to the
index being scanned.

In the future, this option will cause the columns provided by this
relation to be a fixed schema, in order to bubble up errors encountered
while scanning.